// packages/agent-coordination/src/types/coordination.types.ts
export enum SerializationFormat {
  JSON = 'json',
  MSGPACK = 'msgpack',
}